EntityGraph/requirements.txt
pandas
numpy
youtube-dl
pytube
pydub
SpeechRecognition 
ffmpeg
langdetect
openai-whisper
YoutubeTags
rdflib
txtai